Revert "The original version has tensorflow problem using Python3, So I used …"

+++ utils.py
@@ -1,4 +1,3 @@
-# -*- coding: UTF-8 -*-
 import cv2 as cv
 import os
 import sys
@@ -52,4 +51,4 @@ def set_video_writer(cap, write_fps=15):
     return cv.VideoWriter(out_file_path,
                           cv.VideoWriter_fourcc(*'mp4v'),
                           write_fps,
-                          (int(round(cap.get(cv.CAP_PROP_FRAME_WIDTH))), int(round(cap.get(cv.CAP_PROP_FRAME_HEIGHT)))))
+                          (round(cap.get(cv.CAP_PROP_FRAME_WIDTH)), round(cap.get(cv.CAP_PROP_FRAME_HEIGHT))))
